mix logic of templates of doi-metrics badge instead of abstraction
