--- app/templates/components/doi-metrics-badge.hbs
@@ -1,21 +1,0 @@
-<div class="metrics-counter">
-  {{#if hasMetrics}}
-      {{#if citations}}
-        <i class="metrics-icons"> {{fa-icon "quote-right"}} {{ format-numbers citations }} {{  pluralize citations 'Citation' without-count=true}}</i>
-      {{else}}
-        <i class="metrics-icons">{{fa-icon "info-circle"}}  No citations were reported.</i>
-      {{/if}}
-      {{#if hasUsage }}
-        {{#if views}}
-          <i class="metrics-icons"> {{fa-icon "eye"}}  {{ format-numbers views }}  {{ pluralize views 'View' without-count=true }} </i>
-        {{/if}}
-        {{#if downloads}}
-          <i class="metrics-icons"> {{fa-icon "download"}}  {{ format-numbers downloads }}   {{ pluralize downloads 'Download' without-count=true }} </i>
-        {{/if}}
-      {{else}}
-        <i class="metrics-icons">{{fa-icon "info-circle"}}  No usage information was reported.</i>
-      {{/if}}
-  {{else}}
-    <i>{{fa-icon "info-circle"}}  No citations were reported. No usage information was reported.</i>
-  {{/if}}
-</div>
